Revert "Update signup form labels and default value"
This reverts commit fea97679b5e08ee5692b2232dd1157d00d9e90ab.

The previous commit was incomplete in that the UI represented the inverted
logic, but the usage of that value was not inverted as well. Long story short,
Ops Manager currently does not seem to support inverting values; and even if
that's added in the future, we'll stick with the positive case so we don't
have to worry about backwards compatibility with Ops Manager.