Change rep job to use new trusted-certs property
Analogue of https://github.com/pivotal-cf/p-runtime/pull/292.

Now that PAS 2.1 consumes diego-release v1.34.0, it can configure the rep job to use the `containers.trusted_ca_certificates` property that deprecates the previous `diego.rep.trusted_certs` property. The deprecated property will be removed in the forthcoming Diego v2 release line.

More information about this property is available at https://github.com/cloudfoundry/diego-release/blob/v1.34.0/jobs/rep/spec#L290-L307, but it is designed to take a list of strings, each of which accepts a concatenation of CA certificates (which makes it suitable to accept `$ops_manager.trusted_certificates` as a single item). The cell rep presents these certificates as individual files to containers in a documented location (`/etc/cf-system-certificates`).

For illustration, usage of this property in cf-deployment:
- https://github.com/cloudfoundry/cf-deployment/blob/v1.12.0/operations/experimental/enable-instance-identity-credentials.yml#L8-L10
- https://github.com/cloudfoundry/cf-deployment/blob/v1.12.0/operations/experimental/secure-service-credentials.yml#L73-L75

Note that this change should **not** apply to previous ERT or PAS versions.