remove consul agents
There's still a reference to the consul_agent property that provides a
cert. This is because the consul server must still exist for all other
tiles that have not yet been forced to remove their consul agents.

To elaborate:

The agent cert is distributed to all consul_agents by the main server
being configured with the agent_cert and agent_key, and providing a
"consul_client" link, which is consumed by all of the agents.

Even though we no longer have agents in this tile which depend on this
provided link, we cannot remove the property because there might be
other tiles which depend on this cross-deployment link being available.

[#159177064]

Signed-off-by: Michelle He <mhe@pivotal.io>