Remove consul agents
It looks like we also still configured consul in the diego rep, even
though we should have migrated to SQL locks a long time ago.

We checked p-runtime and it's been a long time since we configured
consul in the diego rep there, so we're mostly confident this can be
removed.

[#159177064] remove consul agents

Signed-off-by: Chris Hendrix <chendrix@pivotal.io>