enable nats-tls job on ert and srt
Introduces new nats-tls job colocated with nats job. This job is to
support TLS between all components on PAS. We need to provide a manual
link to allow us to modify the instances to provide the correct nats dns
names to jobs instead of the `.bosh` dns or the ip addresses. This dns
name is required to match certs we are generating.

We intend to improve this configuration in the future.

The existing nats job is able to talk to the nats-tls job. Because of
this, not all jobs need to move to TLS immediately. They will continue
to recieve all messages regardless of which nats server they publish and
subscribe to.

[#170836335](https://www.pivotaltracker.com/story/show/170836335)

Co-authored-by: Aidan Obley <aobley@pivotal.io>
Co-authored-by: Eli Wrenn <ewrenn@pivotal.io>
Co-authored-by: Utako Ueda <uueda@pivotal.io>