Remove istio routers (service mesh)
This was an experimental feature that we do not intend to GA.

The form fields were removed in patch releases of earlier versions of
TAS, but we kept the property and jobs in order to avoid making a
breaking change in patch releases.

A migration ensures that istio is disabled prior to staging.

This removes the feature completely for TAS 2.11+.

Co-authored-by: Amelia Downs <adowns@vmware.com>
Co-authored-by: Josh Russett <jrussett@vmware.com>

[#173738292] Deprecate Dashboard